backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
google-generativeai==0.8.3
manim==0.18.1
moviepy==1.0.3
requests==2.32.3
python-multipart==0.0.12
Pillow==10.4.0
